add Ceil component
