+++ src/components/Ceil/Ceil.module.css.d.ts
@@ -1,10 +1,10 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
+  'bright': string;
+  'footer': string;
   'matrixCeil': string;
-  'matrixRow': string;
-  'sidebarRow': string;
-  'sum': string;
+  'part': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
