+++ src/config/config.tsx
@@ -1,3 +1,3 @@
-export const M = 100; // кількість стрічок
-export const N = 100; //кількість стовпчиків
+export const M = 10; // кількість стрічок
+export const N = 10; //кількість стовпчиків
 export const X = 7; // кількість близьких Amount
